lib: Makefile: allow '~' chars in the uname output
Since we use '~' as our sed delimeter we run into issues if there is a
'~' found in the output of 'uname -a'. We could potentially use a
different delimeter but that just shifts the problem to a different
character. We therefor escape the '~' characters to allow them to
survive the sed expression without causing an error.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>